backend/requirements.txt
mediapipe>=0.10.9
opencv-python>=4.8.0
numpy>=1.24.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
openai-whisper>=20250625
openai>=1.0.0
pypdf>=4.0.0
